test: extract shared OneSlotStore fake into store_fakes
